+++ pkg/database/testdata/phpmyadmin-backup.sql
@@ -1,0 +1,50 @@
+-- phpMyAdmin SQL Dump
+-- version 4.9.5
+-- https://www.phpmyadmin.net/
+--
+-- Host: localhost:3306
+-- Generation Time: Mar 02, 2021 at 05:48 PM
+-- Server version: 5.7.24
+-- PHP Version: 7.3.7
+
+SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
+SET AUTOCOMMIT = 0;
+START TRANSACTION;
+SET time_zone = "+00:00";
+
+
+/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
+/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
+/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
+/*!40101 SET NAMES utf8mb4 */;
+
+--
+-- Database: `nitrotest_craft`
+--
+
+-- --------------------------------------------------------
+
+--
+-- Table structure for table `assetindexdata`
+--
+
+CREATE TABLE `assetindexdata` (
+  `id` int(11) NOT NULL,
+  `sessionId` varchar(36) NOT NULL DEFAULT '',
+  `volumeId` int(11) NOT NULL,
+  `uri` text,
+  `size` bigint(20) UNSIGNED DEFAULT NULL,
+  `timestamp` datetime DEFAULT NULL,
+  `recordId` int(11) DEFAULT NULL,
+  `inProgress` tinyint(1) DEFAULT '0',
+  `completed` tinyint(1) DEFAULT '0',
+  `dateCreated` datetime NOT NULL,
+  `dateUpdated` datetime NOT NULL,
+  `uid` char(36) NOT NULL DEFAULT '0'
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+--
+-- Dumping data for table `assetindexdata`
+--
+
+INSERT INTO `assetindexdata` (`id`, `sessionId`, `volumeId`, `uri`, `size`, `timestamp`, `recordId`, `inProgress`, `completed`, `dateCreated`, `dateUpdated`, `uid`) VALUES